+++ packages/fiori/src/NavigationMenuItem.ts
@@ -2,8 +2,11 @@ import customElement from "@ui5/webcomponents-base/dist/decorators/customElement
 import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import type { ClassMap } from "@ui5/webcomponents-base/dist/types.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
-import MenuItem from "./MenuItem.js";
+import MenuItem from "@ui5/webcomponents/dist/MenuItem.js";
+import type SideNavigationItemDesign from "./types/SideNavigationItemDesign.js";
 import NavigationMenu from "./NavigationMenu.js";
+
+// Templates
 import NavigationMenuItemTemplate from "./NavigationMenuItemTemplate.js";
 
 // Styles
@@ -29,7 +32,7 @@ import {
  *
  * ### ES6 Module Import
  *
- * `import "@ui5/webcomponents/dist/NavigationMenuItem.js";`
+ * `import "@ui5/webcomponents-fiori/dist/NavigationMenuItem.js";`
  * @constructor
  * @extends MenuItem
  * @since 1.22.0
@@ -73,6 +76,9 @@ class NavigationMenuItem extends MenuItem {
 	@property()
 	target?: string;
 
+	@property()
+	design: `${SideNavigationItemDesign}` = "Default";
+
 	get isExternalLink() {
 		return this.href && this.target === "_blank";
 	}
@@ -82,11 +88,15 @@ class NavigationMenuItem extends MenuItem {
 	}
 
 	get _accInfo() {
-		const accInfoSettings = {
-			role: this.href ? "none" as const : "treeitem" as const,
-		};
+		const accInfo = super._accInfo;
 
-		return { ...super._accInfo, ...accInfoSettings };
+		accInfo.role = this.href ? "none" : "treeitem";
+
+		if (!accInfo.ariaHaspopup) {
+			accInfo.ariaHaspopup = this.accessibilityAttributes.hasPopup;
+		}
+
+		return accInfo;
 	}
 
 	get classes(): ClassMap {